+++ main_app/views.py
@@ -4,7 +4,7 @@ from django.contrib.auth import login
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.views.generic import ListView, DetailView
-from .forms import RegistrationForm, CommentForm, AvatarForm
+from .forms import RegistrationForm, CommentForm, CommentUpdateForm, AvatarForm
 from .models import Global, Country, Province, Comment, Profile, County
 from .utils import delete_file
 
@@ -15,18 +15,15 @@ def home(request):
 
 
 def signup(request):
-    error_message = ''
     if request.method == 'POST':
         form = RegistrationForm(request.POST, request.FILES)
         if form.is_valid():
             user = form.save()
             login(request, user)
             return redirect('home')
-        else:
-            error_message = 'Invalid sign up - try again'
-    form = RegistrationForm()
-    context = {'form': form, 'error_message': error_message}
-    return render(request, 'registration/signup.html', context)
+    else:
+        form = RegistrationForm()
+    return render(request, 'registration/signup.html', {'form': form})
 
 
 @login_required
@@ -53,27 +50,28 @@ def delete_comment(request, country_id, comment_id):
 def update_comment(request, country_id, comment_id):
     comment = Comment.objects.get(id=comment_id)
     if request.user.id == comment.created_by.id:
-        comment.content = request.POST.get('content')
-        comment.save()
+        form = CommentUpdateForm(request.POST)
+        if form.is_valid():
+            comment.content = form.save()
+            comment.save()
     return redirect('countries_detail', pk=country_id)
 
 
 def profiles_detail(request, user_id):
+    if request.method == 'POST':
+        form = AvatarForm(request.POST, request.FILES)
+        if form.is_valid():
+            profile = Profile.objects.get(user_id=request.user.id)
+            if profile.avatar:
+                delete_file(profile.avatar)
+            profile.avatar = form.save()
+            profile.save()
+            return redirect('profiles_detail', user_id=request.user.id)
+    else:
+        form = AvatarForm()
     user = User.objects.get(id=user_id)
-    avatar_form = AvatarForm()
-    return render(request, 'main_app/profile_detail.html', {'profile_user': user, 'avatar_form': avatar_form})
-
-
-@login_required
-def update_avatar(request):
-    form = AvatarForm(request.POST, request.FILES)
-    if form.is_valid():
-        profile = Profile.objects.get(user_id=request.user.id)
-        if profile.avatar:
-            delete_file(profile.avatar)
-        profile.avatar = form.save()
-        profile.save()
-    return redirect('profiles_detail', user_id=request.user.id)
+    context = {'profile_user': user, 'avatar_form': form}
+    return render(request, 'main_app/profile_detail.html', context)
 
 
 @login_required
